requirements.txt
flask==2.3.3
tensorflow>=2.15

pillow==10.0.1
numpy==1.24.3
werkzeug==2.3.7 